Compartimentation du generatedContent et checkpoint avant d'appliquer shortDescription sur les autres entités
