+++ src/GeneratedContent/GenerateContent.ts
@@ -1,0 +1,9 @@
+import { generateShortDescriptionFromDescription } from "./ShortDescription/GenerateShortDescription";
+
+export function generateEntityContent(document: any) {
+    let generatedContent = {
+        shortDescription: generateShortDescriptionFromDescription(document?.description),
+    };
+
+    return generatedContent;
+}
